+++ library/src/main/java/com/vorlonsoft/android/rate/DialogManager.java
@@ -11,7 +11,6 @@ import android.content.Context;
 
 public interface DialogManager {
 
-
     interface Factory {
         DialogManager createDialogManager(final Context context, final DialogOptions options);
     }